Move flickrSearch and tagClassifier to utils folder
