+++ server/placeSearch/placeSearchController.js
@@ -3,7 +3,7 @@ const rp = require('request-promise');
 const Promise = require('bluebird');
 const _ = require('lodash');
 const helpers = require('../config/helpers');
-const tagClassifier = require('./tagClassifier');
+const tagClassifier = require('../utils/tagClassifier');
 const GOOGLE_PLACES_URL = 'https://maps.googleapis.com/maps/api/place/textsearch/json';
 const MAX_PLACES_RETURNED = 20;
 const redis = require('redis');
